move to price/prices.utils.ts
